refactor: split naming rule dialog into component + hook, move out of menus folder [DA-419]
- NamingRuleDialogContent manages its own state and actions; no more ref passing
- Hook just opens the dialog; all logic lives in the content component
- Roll back extraAction slot on DialogConfig (no longer needed)
- Folder path now a proper field on FolderTreeItem instead of extracted from ID
- Tone down matching strategy debug logging
