+++ packages/danmaku-anywhere/src/common/components/Dialog/dialogStore.ts
@@ -20,8 +20,6 @@ export interface DialogConfig {
   showCloseButton?: boolean
   closeOnError?: boolean
   container?: HTMLElement | null
-  /** Extra element rendered at the start of the action bar (left side) */
-  extraAction?: ReactNode
 }
 
 type DialogInput = Omit<DialogConfig, 'id'>
