+++ packages/danmaku-anywhere/src/common/localization/locales/en/translation.json
@@ -376,6 +376,7 @@
   },
   "namingRule": {
     "create": "Create Naming Rule",
+    "deleteConfirm": "Are you sure you want to delete this naming rule?",
     "duplicateTitle": "A naming rule with this title already exists",
     "edit": "Edit Naming Rule",
     "episodeLabel": "Episode {{number}}",
@@ -386,6 +387,7 @@
     "noRules": "No naming rules defined",
     "rootFolder": "(root)",
     "ruleCreated": "Naming rule saved",
+    "ruleCreateFailed": "Failed to save naming rule",
     "ruleDeleted": "Naming rule deleted",
     "template": "Filename Template",
     "templateHelp": "Use {episode} for episode number, {episode:02d} for zero-padded",
